Fold ContSpec into IOSpec
